refactor(rstest): drop the puppeteer provider and move the API to the root entry
Test-runner usage has consolidated on Playwright (the Puppeteer core team
moved to Microsoft to build it years ago, and playwright now outdownloads
puppeteer >6x), so ship a single provider built on @rstest/playwright and
keep the surface minimal:

- the root export now provides test/expect/hooks/defineMidsceneDefaults;
  the ./playwright and ./puppeteer subpaths are gone
- @rstest/playwright becomes a regular dependency (implementation detail),
  playwright stays a required peer
- a future @rstest/puppeteer can be added back as a new subpath without
  breaking the root API
