+++ packages/rstest/README.md
@@ -1,5 +1,5 @@
 # @midscene/rstest
 
-Run Midscene AI browser agents as [Rstest](https://rstest.rsbuild.dev/) tests. Provides a Playwright provider built on [`@rstest/playwright`](https://www.npmjs.com/package/@rstest/playwright) (inheriting its debug mode, trace capture, and Playwright-flavored `expect`), a Puppeteer provider, and an Rstest reporter that surfaces Midscene reports.
+Run Midscene AI browser agents as [Rstest](https://rstest.rsbuild.dev/) tests. Built on [`@rstest/playwright`](https://www.npmjs.com/package/@rstest/playwright) — inheriting its debug mode, trace capture, and Playwright-flavored `expect` — plus an Rstest reporter that surfaces Midscene reports.
 
 - **Integration guide**: <https://midscenejs.com/integrate-with-rstest.html>
